Refactor new connection handling
Removes PhpDebugSession.prototype._mainConnection. The first connection
is not treated differently than following connections anymore and
breakpoints are not stored anymore. Instead, on each new connection
an InitializedEvent is triggered. Connections that have not yet been
initialized with breakpoints are stored in a Set, which is checked in
both setBreakpointRequest and setExceptionBreakpointRequest. This fixes
a race condition between the two which could result in a run command
before all breakpoints were set. The two Sets are a workaround until
ConfigurationDoneRequest ships in VS Code.