fix: reset "stopOnEntry" after first use to show correct pause reason (#478)
* Reset "stopOnEntry" after first use to show correct pause reason.

If "stopOnEntry" is enabled the pause reason will always be "Paused on entry" event if it's the result of a step operation or breakpoint. Resetting the "stopOnEntry" after first use solves the issue. The flag is not used anywhere else.

* Do not use _args.stopOnEntry since it should be regarded immutable and is set only on every adapter run.
Reconnecting to adapter would cause issues.
Always reset _hasStoppedOnEntry flag just before calling the step into operation.